QuickPay: Fix re-using stored cards
A token needs to be generated for each operation against a stored card. Previously a token was created when a card was stored and then it was not possible to generate any more. The card storage and usage of stored cards has changed to automatically generate tokens when needed.

NOTE: The return types of a few methods have a changed a bit and been simplified. Most notablly, storing a card now only returns the card ID and not a token. Exposing the token to the implementer can only serve to create confusion and problems